+++ ChangeLog.md
@@ -1,4 +1,8 @@
-# 1.0.1 -- 2017-02-06
+# 1.0.1 -- 2017-05-05
+
+* Add support for `x-doctest-options` cabal-file field
+
+* Proper support for GHC-8.2.1 & Cabal-2.0.0.0
 
 * Add support to `default-extensions` in library.